refactor: Add GetUser, RawHeaders decorators and update AuthController
This commit refactors the AuthController in the 04-e-comerce project. It adds the GetUser and RawHeaders decorators to retrieve user information from the request and access raw headers, respectively. The GetUser decorator is used to get the user object from the request, while the RawHeaders decorator is used to access the raw headers of the HTTP request.

These changes improve the functionality and flexibility of the AuthController by providing additional information and access to the user and headers. This will be useful for implementing authentication and authorization features in the application.

This commit also includes updates to the import statements and the constructor of the AuthController to reflect the changes made.

Ref: #123