refactor: Update ProductsController and ProductsService to include user information in create and update methods
This commit updates the ProductsController and ProductsService in the 04-e-comerce project to include user information in the create and update methods. The @GetUser decorator is used to retrieve the user object from the request, and it is passed as an argument to the create and update methods.

This change improves the functionality and flexibility of the application by associating the created or updated products with the user who performed the action. It provides better tracking and accountability for product changes.

Ref: #123